extracts keydown listener to moveWithArrowKeys
